modified labels removed unwanted options

+++ stp/student_controller_update_task.php
@@ -14,7 +14,14 @@ $hours = $_POST['time'];
 $user_id=$_POST['user_id'];
 $task_id=$_POST['task_id'];
 $completed_hour=$_POST['completed_hour'];
+$completed_hours=$hours;
+$preferred_hour=$_POST['preferred_hour'];
 $hours=$hours+$completed_hour;
+//check if competed hour grater than preffered hour
+if($hours>$preferred_hour){
+    header('Location: student_tasks.php?hour=false');
+    exit();
+}
 //delete row
 $delete_query="DELETE from tbl_student_task where tbl_student_task.task_id=? and tbl_student_task.student_id=?";
 if(($stmt=$connection->prepare($delete_query)))
